[BUGFIX] Normalize public URLs by removing leading slash
Unfortunately the response from `getPath()` in TYPO3 is not consistent.
Sometimes it contains a leading slash sometimes it doesn't.

There was already one occurrence where leading slashes were removed.

This change introduces a simple static method
`PathUtility::stripLeadingSlash()` which is then used everywhere a path is
read from or written to the database.